+++ CHANGELOG.md
@@ -5,6 +5,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+### Changed
+- Pre-publish and publish timer beans run for a given time slot and will stop before the next timer starts
+
 ## [1.1.9] 2019-05-10
 ### Added
 - Retrofit the throttling of event publishing to framework 5